beolvasas

+++ 00_Valtozok/solution_valtozo/feladat_02/Program.cs
@@ -1,5 +1,5 @@
 ﻿string modell = "i7";
 int type = 8700;
-float power = 4.6;
+double power = 4.6;
 
 Console.WriteLine($"Az INTEL {modell} sorozatjelzésű mikroprocesszor {type} típusának munkafrekvenciája {power}GHz");
